+++ FrontEnd/src/pages/Logout/Logout.jsx
@@ -1,9 +1,9 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
-import { clearToken } from "../../redux/features/token.js";
-import { clearFirstName } from "../../redux/features/firstName.js";
-import { clearLastName } from "../../redux/features/lastName.js";
+import { clearToken } from "../../redux/reducers.js";
+import { clearFirstName } from "../../redux/reducers.js";
+import { clearLastName } from "../../redux/reducers.js";
 
 export function Logout() {
     const dispatch = useDispatch();
