+++ FrontEnd/src/pages/Profil/Profil.jsx
@@ -2,9 +2,9 @@ import './profil.scss';
 import { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Navigate } from 'react-router-dom';
-import { getLoginFetch, saveUserProfile } from '../../utils/api.js';
-import { setFirstName } from "../../redux/features/firstName.js";
-import { setLastName } from "../../redux/features/lastName.js";
+import { getLoginFetch, saveUserProfile } from '../../utils/api/user.js';
+import { setFirstName } from "../../redux/reducers.js";
+import { setLastName } from "../../redux/reducers.js";
 import { selectToken, selectFirstName, selectLastName } from '../../redux/selectors.js';
 import Account from '../../components/Account/Account.js';
 import accountsMocks from '../../mocks/accountsMocks.js';
